move function to convert KeymanagerOpts to Config inside rpc and run gazelle
